requirements.txt

mlflow
scikit-learn
pandas
requests
matplotlib
seaborn
sqlalchemy
fastapi
uvicorn
pydantic
numpy
pyyaml
pytest
